Added port list to cr16 (see description)
I realize that this port list would actually just be a set of variables if cr16 is going to be our top-level design. I am having a hard time conceptualizing how this Verilog file will look without a separate FSM control module that interfaces with BRAM, the datapath, and the program counter.